+++ src/main/steps/applicant1Sequence.ts
@@ -38,7 +38,7 @@ import {
   HOW_DO_YOU_WANT_TO_APPLY,
   HOW_THE_COURTS_WILL_CONTACT_YOU,
   HOW_TO_APPLY_TO_SERVE,
-  HUB_1,
+  HUB_PAGE,
   IN_THE_UK,
   JURISDICTION_CONNECTION_SUMMARY,
   JURISDICTION_DOMICILE,
@@ -435,7 +435,7 @@ export const applicant1Sequence: Step[] = [
     getNextStep: () => HOME_URL,
   },
   {
-    url: HUB_1,
+    url: HUB_PAGE,
     getNextStep: () => HOME_URL,
   },
 ];
