--- src/main/steps/applicant2/hub-1/template.njk
@@ -1,1 +1,0 @@
-{% include "../../applicant1/hub-1/template.njk" %}
