+++ src/main/steps/applicant2/hub-page/content.ts
@@ -1,6 +1,9 @@
 import { TranslationFn } from '../../../app/controller/GetController';
 import { FormContent } from '../../../app/form/Form';
-import { form as applicant1Form, generateContent as applicant1GenerateContent } from '../../applicant1/hub-1/content';
+import {
+  form as applicant1Form,
+  generateContent as applicant1GenerateContent,
+} from '../../applicant1/hub-page/content';
 
 export const form: FormContent = {
   ...applicant1Form,
