+++ src/main/steps/applicant2/hub-page/template.njk
@@ -1,0 +1,1 @@
+{% include "../../applicant1/hub-page/template.njk" %}
