+++ src/main/steps/urls.ts
@@ -99,5 +99,5 @@ export const EQUALITY: PageLink = '/equality';
 export const PAYMENT_CALLBACK_URL: PageLink = '/payment-callback';
 
 export const CONFIRM_JOINT_APPLICATION: PageLink = '/confirm-your-joint-application';
-export const HUB_1: PageLink = '/hub-1';
+export const HUB_PAGE: PageLink = '/hub-page';
 export const REVIEW_CONTACT_DETAILS: PageLink = '/check-contact-details';
